requirements.txt
Flask==2.2.3
PyJWT==2.6.0
pymongo==4.3.3
xgboost==1.7.5
requests==2.28.2
python-dotenv==1.0.0
SpeechRecognition==3.8.1
pydub==0.25.1
bcrypt==4.0.1
werkzeug==2.2.3
